update common and update to follow best practices

+++ techlock/user_management_service/routes/roles.py
@@ -1,18 +1,18 @@
 import logging
-from typing import List
+from dataclasses import asdict
+from typing import Any, Dict
 from uuid import UUID
 
 from flask.views import MethodView
 from flask_smorest import Blueprint
 from techlock.common.api import BadRequestException, Claim, NotFoundException
-from techlock.common.api.auth import (
-    access_required,
-    get_current_user_with_claims,
-)
+from techlock.common.api.auth import access_required
+from techlock.common.api.auth.claim import ClaimSet
+from techlock.common.api.models.dry_run import DryRunSchema
 from techlock.common.config import AuthInfo
 
+from ..models import ROLE_CLAIM_SPEC as claim_spec
 from ..models import (
-    ROLE_CLAIM_SPEC,
     Role,
     RoleListQueryParameters,
     RoleListQueryParametersSchema,
@@ -34,7 +34,7 @@ def set_claims_default_tenant(data: dict, default_tenant_id: UUID):
             for claim in claims:
                 c = Claim.from_string(claim)
                 if c.tenant_id == '':
-                    c.tenant_id = default_tenant_id
+                    c = Claim(**{**asdict(c), 'tenant_id': default_tenant_id})
                 new_claims = new_claims + [str(c)]
             claims_by_audience[key] = new_claims
     return claims_by_audience
@@ -47,15 +47,10 @@ class Roles(MethodView):
         MethodView.__init__(self, *args, **kwargs)
         self.idp = get_idp()
 
+    @access_required('read', claim_spec=claim_spec)
     @blp.arguments(schema=RoleListQueryParametersSchema, location='query')
     @blp.response(status_code=200, schema=RolePageableSchema)
-    @access_required(
-        'read', 'roles',
-        allowed_filter_fields=ROLE_CLAIM_SPEC.filter_fields
-    )
-    def get(self, query_params: RoleListQueryParameters):
-        current_user, claims = get_current_user_with_claims()
-
+    def get(self, query_params: RoleListQueryParameters, current_user: AuthInfo, claims: ClaimSet):
         pageable_resp = Role.get_all(
             current_user,
             offset=query_params.offset,
@@ -65,26 +60,25 @@ class Roles(MethodView):
             claims=claims,
         )
 
-        logger.info('GET roles', extra={
-            'roles': pageable_resp.asdict()
-        })
+        logger.info('GET roles', extra={'roles': pageable_resp.asdict()})
 
         return pageable_resp
 
+    @access_required('create', claim_spec=claim_spec)
     @blp.arguments(schema=RoleSchema)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
     @blp.response(status_code=201, schema=RoleSchema)
-    @access_required('create', 'roles')
-    def post(self, data):
-        current_user, claims = get_current_user_with_claims()
-        logger.info('Creating Role', extra={'data': data})
+    def post(self, data: Dict[str, Any], dry_run: bool, current_user: AuthInfo, claims: ClaimSet):
+        logger.info('Creating role', extra={'data': data})
 
-        # Role.validate(data)
         role = Role(**data)
         role.claims_by_audience = set_claims_default_tenant(data, current_user.tenant_id)
 
-        role.save(current_user, claims=claims)
+        # no need to rollback on dry-run, flask-sqlalchemy does this for us.
+        role.save(current_user, claims=claims, commit=not dry_run)
 
-        self.idp.update_or_create_role(current_user, role)
+        if not dry_run:
+            self.idp.update_or_create_role(current_user, role)
 
         return role
 
@@ -96,61 +90,58 @@ class RoleById(MethodView):
         MethodView.__init__(self, *args, **kwargs)
         self.idp = get_idp()
 
-    def get_role(self, current_user: AuthInfo, claims: List[Claim], role_id: str):
+    def get_role(self, current_user: AuthInfo, claims: ClaimSet, role_id: str):
         role = Role.get(current_user, role_id, claims=claims, raise_if_not_found=True)
 
         return role
 
+    @access_required('read', claim_spec=claim_spec)
     @blp.response(status_code=200, schema=RoleSchema)
-    @access_required(
-        'read', 'roles',
-        allowed_filter_fields=ROLE_CLAIM_SPEC.filter_fields
-    )
-    def get(self, role_id):
-        current_user, claims = get_current_user_with_claims()
-
+    def get(self, role_id: str, current_user: AuthInfo, claims: ClaimSet):
+        logger.info('Getting role', extra={'id': role_id})
         role = self.get_role(current_user, claims, role_id)
 
         return role
 
+    @access_required(['read', 'update'], claim_spec=claim_spec)
     @blp.arguments(schema=RoleSchema)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
     @blp.response(status_code=200, schema=RoleSchema)
-    @access_required(
-        'update', 'roles',
-        allowed_filter_fields=ROLE_CLAIM_SPEC.filter_fields
-    )
-    def put(self, data, role_id):
-        current_user, claims = get_current_user_with_claims()
-        logger.debug('Updating Role', extra={'data': data})
+    def put(self, data: Dict[str, Any], dry_run: bool, role_id: str, current_user: AuthInfo, claims: ClaimSet):
+        logger.debug('Updating role', extra={'data': data})
 
-        role = self.get_role(current_user, claims, role_id)
+        role = self.get_role(current_user, claims.filter_by_action('read'), role_id)
 
         for k, v in data.items():
             if hasattr(role, k):
                 setattr(role, k, v)
             else:
-                raise BadRequestException('Role has no attribute: %s' % k)
+                raise BadRequestException(f'Role has no attribute: {k}')
 
         role.claims_by_audience = set_claims_default_tenant(data, current_user.tenant_id)
 
-        role.save(current_user, claims=claims)
-        self.idp.update_or_create_role(current_user, role)
+        # no need to rollback on dry-run, flask-sqlalchemy does this for us.
+        role.save(current_user, claims=claims.filter_by_action('update'), commit=not dry_run)
+        if not dry_run:
+            self.idp.update_or_create_role(current_user, role)
 
         return role
 
-    @blp.response(status_code=204, schema=RoleSchema)
-    @access_required(
-        'delete', 'roles',
-        allowed_filter_fields=ROLE_CLAIM_SPEC.filter_fields
-    )
-    def delete(self, role_id):
-        current_user, claims = get_current_user_with_claims()
-        role = self.get_role(current_user, claims, role_id)
+    @access_required(['read', 'delete'], claim_spec=claim_spec)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
+    @blp.response(status_code=204)
+    def delete(self, dry_run: bool, role_id: str, current_user: AuthInfo, claims: ClaimSet):
+        logger.info('Deleting role', extra={'id': role_id})
 
-        try:
-            self.idp.delete_role(current_user, role)
-        except NotFoundException:
-            logger.warning('Role does not exist in IDP, skipping IDP deletion...', extra={'role_idp_name': role.idp_name})
+        role = self.get_role(current_user, claims.filter_by_action('read'), role_id)
 
-        role.delete(current_user)
-        return role
+        if not dry_run:
+            try:
+                self.idp.delete_role(current_user, role)
+            except NotFoundException:
+                logger.warning('Role does not exist in IDP, skipping IDP deletion...', extra={'role_idp_name': role.idp_name})
+
+        # no need to rollback on dry-run, flask-sqlalchemy does this for us.
+        role.delete(current_user, claims=claims.filter_by_action('delete'), commit=not dry_run)
+
+        return
