+++ techlock/user_management_service/routes/users.py
@@ -1,30 +1,25 @@
 import logging
-from typing import List, Union
+from dataclasses import asdict
+from typing import Any, Dict, List, Tuple, Union
 from uuid import UUID
 
 from flask.views import MethodView
 from flask_smorest import Blueprint
 from sqlalchemy import and_ as sql_and_  # diffentiate from operators.and_
-from techlock.common.api import (
-    BadRequestException,
-    ConflictException,
-    NotFoundException,
-)
-from techlock.common.api.auth import (
-    Claim,
-    access_required,
-    can_access,
-    get_current_user_with_claims,
-)
+from techlock.common.api import BadRequestException, ConflictException
+from techlock.common.api.auth import Claim, access_required
+from techlock.common.api.auth.claim import ClaimSet, ClaimSpec
 from techlock.common.api.auth.utils import SYSTEM_TENANT_ID
+from techlock.common.api.models.dry_run import DryRunSchema
 from techlock.common.config import AuthInfo, ConfigManager
 from techlock.common.messaging.sns import Envelope
 from techlock.common.messaging.sns import publish as publish_sns
 from techlock.common.orm.sqlalchemy import BaseModel
 from techlock.common.util.helper import suppress_with_log
 
+from ..models import DEPARTMENT_CLAIM_SPEC, OFFICE_CLAIM_SPEC, ROLE_CLAIM_SPEC
+from ..models import USER_CLAIM_SPEC as claim_spec
 from ..models import (
-    USER_CLAIM_SPEC,
     Department,
     Office,
     PostUserChangePasswordSchema,
@@ -43,25 +38,49 @@ logger = logging.getLogger(__name__)
 
 blp = Blueprint('users', __name__, url_prefix='/users')
 
-idp_attribute_keys = [
-]
+idp_attribute_keys = []
+
+merged_claim_spec = ClaimSpec(
+    resource_name=[
+        claim_spec.resource_name,
+        DEPARTMENT_CLAIM_SPEC.resource_name,
+        OFFICE_CLAIM_SPEC.resource_name,
+        ROLE_CLAIM_SPEC.resource_name,
+    ],
+    filter_fields=list(set(
+        claim_spec.filter_fields
+        + DEPARTMENT_CLAIM_SPEC.filter_fields
+        + OFFICE_CLAIM_SPEC.filter_fields
+        + ROLE_CLAIM_SPEC.filter_fields  # noqa: C812
+    )),
+    default_actions=list(set(
+        claim_spec.default_actions
+        + DEPARTMENT_CLAIM_SPEC.default_actions
+        + OFFICE_CLAIM_SPEC.default_actions
+        + ROLE_CLAIM_SPEC.default_actions  # noqa: C812
+    )),
+)
 
 
-def _get_items_from_id_list(current_user: AuthInfo, claims: List[Claim], id_list: List[Union[str, UUID]], orm_class: BaseModel):
-    items = list()
+def _get_items_from_id_list(
+    current_user: AuthInfo,
+    claims: List[Claim],
+    id_list: List[Union[str, UUID]],
+    orm_class: BaseModel,
+):
     if not id_list:
-        return items
+        return []
 
-    for entity_id in id_list:
-        items.append(orm_class.get(current_user, entity_id=entity_id, claims=claims, raise_if_not_found=True))
+    items = [
+        orm_class.get(current_user, entity_id=entity_id, claims=claims, raise_if_not_found=True)
+        for entity_id in id_list
+    ]
 
     return items
 
 
-def _get_user(current_user: AuthInfo, claims: List[Claim], user_id: str):
-    user = User.get(current_user, user_id, claims=claims)
-    if user is None or (current_user.user_id != user_id and not can_access(user, claims)):
-        raise NotFoundException('No user found for id = {}'.format(user_id))
+def _get_user(current_user: AuthInfo, claims: ClaimSet, user_id: str) -> User:
+    user = User.get(current_user, user_id, claims=claims, raise_if_not_found=True)
 
     return user
 
@@ -84,10 +103,31 @@ def set_claims_default_tenant(data: dict, default_tenant_id: UUID):
             for claim in claims:
                 c = Claim.from_string(claim)
                 if c.tenant_id == '':
-                    c.tenant_id = default_tenant_id
+                    c = Claim(**{**asdict(c), 'tenant_id': default_tenant_id})
                 new_claims = new_claims + [str(c)]
             claims_by_audience[key] = new_claims
     return claims_by_audience
+
+
+def _split_claims(claims: ClaimSet, *office_actions: List[str]) -> Tuple[ClaimSet, ClaimSet, ClaimSet, ClaimSet]:
+    user_claims = ClaimSet([
+        c for c in claims
+        if c.action in ('*', *office_actions) and c.resource in ('*', claim_spec.resource_name)
+    ])
+    department_claims = ClaimSet([
+        c for c in claims
+        if c.action in ('*', 'read') and c.resource in ('*', DEPARTMENT_CLAIM_SPEC.resource_name)
+    ])
+    office_claims = ClaimSet([
+        c for c in claims
+        if c.action in ('*', 'read') and c.resource in ('*', OFFICE_CLAIM_SPEC.resource_name)
+    ])
+    role_claims = ClaimSet([
+        c for c in claims
+        if c.action in ('*', 'read') and c.resource in ('*', ROLE_CLAIM_SPEC.resource_name)
+    ])
+
+    return user_claims, department_claims, office_claims, role_claims
 
 
 @blp.route('')
@@ -97,14 +137,10 @@ class Users(MethodView):
         MethodView.__init__(self, *args, **kwargs)
         self.idp = get_idp()
 
+    @access_required('read', claim_spec=claim_spec)
     @blp.arguments(schema=UserListQueryParametersSchema, location='query')
     @blp.response(status_code=200, schema=UserPageableSchema)
-    @access_required(
-        'read', 'users',
-        allowed_filter_fields=USER_CLAIM_SPEC.filter_fields
-    )
-    def get(self, query_params: UserListQueryParameters):
-        current_user, claims = get_current_user_with_claims()
+    def get(self, query_params: UserListQueryParameters, current_user: AuthInfo, claims: ClaimSet):
         if not claims:
             # if no claims, add one that allows the user to see himself
             claims = [Claim(True, current_user.tenant_id, '*', '*', 'users', id=current_user.user_id, filter_field=None, filter_value=None)]
@@ -118,24 +154,21 @@ class Users(MethodView):
             claims=claims,
         )
 
-        logger.info('GET users', extra={
-            'users': pageable_resp.asdict()
-        })
+        logger.info('GET users', extra={'users': pageable_resp.asdict()})
 
         return pageable_resp
 
+    @access_required(['read', 'create'], claim_spec=merged_claim_spec)
     @blp.arguments(schema=PostUserSchema)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
     @blp.response(status_code=201, schema=UserSchema)
-    @access_required('create', 'users')
-    def post(self, data: dict):
-        current_user, claims = get_current_user_with_claims()
-        logger.info('Creating User', extra={'data': data})
+    def post(self, data: Dict[str, Any], dry_run: bool, current_user: AuthInfo, claims: ClaimSet):
+        logger.info('Creating user', extra={'data': data})
         if current_user.tenant_id == SYSTEM_TENANT_ID:
             raise BadRequestException('Can not create system users.')
 
         # Get the password and remove it from the data. It is not part of the User object
         temporary_password = data.pop('temporary_password')
-        # User.validate(data)
 
         # Since users are to be unique across all tenants by email address, we must
         # get the full list of users, even ones from other tenants. Therefore we use
@@ -152,10 +185,11 @@ class Users(MethodView):
             # Validate that ftp_username is unique to the tenant
             raise ConflictException(f"ftp_username '{ftp_username}' already exists.")
 
+        user_claims, department_claims, office_claims, role_claims = _split_claims(claims, 'create')
         # Validate that items exist and get actual items
-        roles = _get_items_from_id_list(current_user, claims=claims, id_list=data.get('role_ids'), orm_class=Role)
-        departments = _get_items_from_id_list(current_user, claims=claims, id_list=data.get('department_ids'), orm_class=Department)
-        offices = _get_items_from_id_list(current_user, claims=claims, id_list=data.get('office_ids'), orm_class=Office)
+        roles = _get_items_from_id_list(current_user, claims=role_claims, id_list=data.get('role_ids'), orm_class=Role)
+        departments = _get_items_from_id_list(current_user, claims=department_claims, id_list=data.get('department_ids'), orm_class=Department)
+        offices = _get_items_from_id_list(current_user, claims=office_claims, id_list=data.get('office_ids'), orm_class=Office)
 
         user = User(
             entity_id=data.get('email'),
@@ -171,28 +205,33 @@ class Users(MethodView):
             offices=offices,
         )
 
-        logger.info('Adding user to idp')
-        idp_attributes = {k: v for k, v in data.items() if k in idp_attribute_keys}
+        if dry_run:
+            logger.info('Dry run, creating user')
+            # Save so that the created_by fields get set
+            user.save(current_user, claims=user_claims, commit=False)
+        else:
+            logger.info('Adding user to idp')
+            idp_attributes = {k: v for k, v in data.items() if k in idp_attribute_keys}
 
-        self.idp.create_user(current_user, user, password=temporary_password, idp_attributes=idp_attributes)
+            self.idp.create_user(current_user, user, password=temporary_password, idp_attributes=idp_attributes)
 
-        try:
-            self.idp.update_user_roles(current_user, user, user.roles)
+            try:
+                self.idp.update_user_roles(current_user, user, user.roles)
 
-            logger.info('User added to idp, storing internally')
-            user.save(current_user, claims=claims)
-            logger.info('User created')
-        except Exception:
-            # Clean up silently. We expect part of this to fail, because if everything was created properly, we wouldn't be here.
-            # Bit of a lazy approach :)
-            with suppress_with_log(logger, Exception, log_level=logging.DEBUG):
-                self.idp.delete_user(current_user, user)
+                logger.info('User added to idp, storing internally')
+                user.save(current_user, claims=user_claims)
+                logger.info('User created')
+            except Exception:
+                # Clean up silently. We expect part of this to fail, because if everything was created properly, we wouldn't be here.
+                # Bit of a lazy approach :)
+                with suppress_with_log(logger, Exception, log_level=logging.DEBUG):
+                    self.idp.delete_user(current_user, user)
 
-            with suppress_with_log(logger, Exception, log_level=logging.DEBUG):
-                user = _get_user(current_user, data['email'])
-                user.delete(current_user)
+                with suppress_with_log(logger, Exception, log_level=logging.DEBUG):
+                    user = _get_user(current_user, data['email'])
+                    user.delete(current_user, claims=user_claims)
 
-            raise
+                raise
 
         return user
 
@@ -204,37 +243,32 @@ class UserById(MethodView):
         MethodView.__init__(self, *args, **kwargs)
         self.idp = get_idp()
 
+    @access_required('read', claim_spec=claim_spec)
     @blp.response(status_code=200, schema=UserSchema)
-    @access_required(
-        'read', 'users',
-        allowed_filter_fields=USER_CLAIM_SPEC.filter_fields
-    )
-    def get(self, user_id: str):
-        current_user, claims = get_current_user_with_claims()
+    def get(self, user_id: str, current_user: AuthInfo, claims: ClaimSet):
+        logger.info('Getting user', extra={'id': user_id})
         user = _get_user(current_user, claims, user_id)
 
         return user
 
+    @access_required(['read', 'update'], claim_spec=merged_claim_spec)
     @blp.arguments(schema=UpdateUserSchema)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
     @blp.response(status_code=200, schema=UserSchema)
-    @access_required(
-        'update', 'users',
-        allowed_filter_fields=USER_CLAIM_SPEC.filter_fields
-    )
-    def put(self, data: dict, user_id: str):
-        current_user, claims = get_current_user_with_claims()
-        logger.debug('Updating User', extra={'data': data})
+    def put(self, data: Dict[str, Any], dry_run: bool, user_id: str, current_user: AuthInfo, claims: ClaimSet):
+        logger.debug('Updating user', extra={'data': data})
+        user_claims, department_claims, office_claims, role_claims = _split_claims(claims, 'create')
 
         # User.validate(data, validate_required_fields=False)
-        user = _get_user(current_user, claims, user_id)
+        user = _get_user(current_user, user_claims.filter_by_action('read'), user_id)
 
         if user.email != data.get('email'):
             raise BadRequestException('Email can not be changed.')
 
         # Validate that items exist and get actual items
-        data['roles'] = _get_items_from_id_list(current_user, claims=claims, id_list=data.pop('role_ids', None), orm_class=Role)
-        data['departments'] = _get_items_from_id_list(current_user, claims=claims, id_list=data.pop('department_ids', None), orm_class=Department)
-        data['offices'] = _get_items_from_id_list(current_user, claims=claims, id_list=data.pop('office_ids', None), orm_class=Office)
+        data['roles'] = _get_items_from_id_list(current_user, claims=role_claims, id_list=data.pop('role_ids', None), orm_class=Role)
+        data['departments'] = _get_items_from_id_list(current_user, claims=department_claims, id_list=data.pop('department_ids', None), orm_class=Department)
+        data['offices'] = _get_items_from_id_list(current_user, claims=office_claims, id_list=data.pop('office_ids', None), orm_class=Office)
 
         ftp_username = data.get('ftp_username')
         if ftp_username == '':
@@ -243,7 +277,7 @@ class UserById(MethodView):
             # Validate that ftp_username is unique to the tenant
             raise ConflictException(f"ftp_username '{ftp_username}' already exists.")
 
-        attributes_to_update = dict()
+        attributes_to_update = {}
         for k, v in data.items():
             if k in idp_attribute_keys:
                 attributes_to_update[k] = v
@@ -252,34 +286,33 @@ class UserById(MethodView):
                 if k in ('name', 'family_name'):
                     attributes_to_update[k] = v
             else:
-                raise BadRequestException('User has no attribute: %s' % k)
+                raise BadRequestException(f'User has no attribute: {k}')
 
         user.claims_by_audience = set_claims_default_tenant(data, current_user.tenant_id)
 
-        user.save(current_user, claims=claims)
+        user.save(current_user, claims=user_claims.filter_by_action('update'), commit=not dry_run)
 
-        self.idp.update_user_attributes(current_user, user, attributes_to_update)
-        self.idp.update_user_roles(current_user, user, data['roles'])
+        if not dry_run:
+            self.idp.update_user_attributes(current_user, user, attributes_to_update)
+            self.idp.update_user_roles(current_user, user, data['roles'])
 
         return user
 
-    @blp.response(status_code=204, schema=UserSchema)
-    @access_required(
-        'delete', 'users',
-        allowed_filter_fields=USER_CLAIM_SPEC.filter_fields
-    )
-    def delete(self, user_id: str):
-        current_user, claims = get_current_user_with_claims()
-
+    @access_required(['read', 'delete'], claim_spec=claim_spec)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
+    @blp.response(status_code=204)
+    def delete(self, dry_run: bool, user_id: str, current_user: AuthInfo, claims: ClaimSet):
         if current_user.user_id == user_id:
             return BadRequestException('Can not delete yourself.')
 
-        user = _get_user(current_user, claims, user_id)
+        user = _get_user(current_user, claims.filter_by_action('read'), user_id)
 
-        self.idp.delete_user(current_user, user)
-        logger.info('Deleted user from userpool')
+        if not dry_run:
+            self.idp.delete_user(current_user, user)
+            logger.info('Deleted user from userpool')
 
-        user.delete(current_user)
+        # no need to rollback on dry-run, flask-sqlalchemy does this for us.
+        user.delete(current_user, claims=claims.filter_by_action('delete'), commit=not dry_run)
         logger.info('Deleted user', extra={'user': user.entity_id})
 
         return
@@ -293,16 +326,18 @@ class UserChangePassword(MethodView):
         self.idp = get_idp()
 
     @blp.arguments(schema=PostUserChangePasswordSchema)
+    @blp.arguments(DryRunSchema, location='query', as_kwargs=True)
     @blp.response(status_code=200)
     @access_required(
         'update', 'user_passwords',
-        allowed_filter_fields=USER_CLAIM_SPEC.filter_fields
+        allowed_filter_fields=claim_spec.filter_fields,
     )
-    def post(self, data: dict, user_id: str):
-        current_user, claims = get_current_user_with_claims()
+    def post(self, data: dict, dry_run: bool, user_id: str, current_user: AuthInfo, claims: ClaimSet):
+        logger.debug('Changing password for user')
         user = _get_user(current_user, claims, user_id)
 
-        self.idp.change_password(current_user, user, data.get('new_password'))
+        if not dry_run:
+            self.idp.change_password(current_user, user, data.get('new_password'))
 
         cm = ConfigManager()
         # Password was successfully changed, log any errors that occur in the post processing
@@ -313,7 +348,8 @@ class UserChangePassword(MethodView):
                 subject='Password Changed',
                 message={
                     'user_id': user_id,
-                    'changed_by': current_user.user_id
+                    'changed_by': current_user.user_id,
+                    'dry_run': dry_run,
                 },
                 tenant_id=current_user.tenant_id,
                 source='user-management-service',
